Fixed Bug in Room 3a & Backgrounds
There was a little bug in room 3a where you couldn't get Octoboy to get in the top space, so I made the spaces a little bigger for him to fit. Also, I made two backgrounds for 3a and 4a. They look quite nice if I do say so myself.